Rename data_source.go to starknetdata.go
