+++ starknetdata/starknetdata.go
@@ -1,11 +1,11 @@
-package datasource
+package starknetdata
 
 import (
 	"github.com/NethermindEth/juno/core"
 	"github.com/NethermindEth/juno/core/felt"
 )
 
-type DataSource interface {
+type StarkNetData interface {
 	GetBlockByNumber(blockNumber uint64) (*core.Block, error)
 	GetTransaction(transactionHash *felt.Felt) (*core.Transaction, error)
 	GetClass(classHash *felt.Felt) (*core.Class, error)
